requirements.txt
streamlit
pandas
plotly
scikit-learn
backboard-sdk